Add CSV unit test file
